Forbid game descriptions with joined islands in Bridges
A game description with islands in adjacent grid squares, like
"3x3:11g", shouldn't be allowed.  If it is, then bridges between the
islands are invisible and clicking one of them causes an assertion
failure: "Assertion `is_loop->adj.points[j].off > 1' failed."

The code to check this is really rather complex, but I think the
complexity is mostly necessary.