js: use the "load" event for loading save files
This is in place of the "loadend" event.  In Chromium, (and in the
specification), "loadend" is triggered not only when the file is
loaded but also when loading fails.  Obviously when loading fails we
don't want to try to parse the (nonexistent) resulting file.

Using the "load" event works better, since it's only fired on success,
and we can also have an "error" handler to report problems with
loading files, albeit with no detail at all.

This doesn't seem to make any difference in Firefox, which in my
testing fires "load" and "loadend" on success and nothing at all on
failure.